update to comment-toggle button
Comment-toggle button now looks like pushed down when toggled on.
Icon on the comment button has been changed so that it reflects the
state of toggle.
Icon for "Raw wiki text" in menu has been changed from "eye-open" to
"file" because it uses the same icon as one for the new comment toggler.
(sorry for changing this!)