fixed glitch when editing SideBar
sidebar() method was added to change the id of wiki-generated sidebar
div box from #sidebar to #sidebar-content (by overriding the original
sidebar() method in __init__.py).
Normally, the id of the div box used to be #SideBar.sidebar. However,
when viewing or editing SideBar page, it was changed to #sidebar which
would conflict with one of the parent div boxes with the same id.